Update README.md
Since Spring now officially supports GraphQL we have decided to archive this project. We encourage you to start using Spring for GraphQL instead. If you're projects rely on this library and you really require an update and cannot migrate to Spring for GraphQL yet, then you should fork this repository and make the necessary changes on your own.